refactor(audio): move client decode types into the waveform types module
