+++ src/lib/viewers/media/waveform/types.ts
@@ -94,6 +94,42 @@ export type WaveformCaps = {
     maxCompressedBytes?: number;
 };
 
+export type DecodeSkipReason = 'missing_metadata' | 'compressed_size' | 'duration';
+
+export type DecodeDecision = { isAllowed: true } | { isAllowed: false; reason: DecodeSkipReason };
+
+export type MediaInfo = {
+    compressedBytes?: number;
+    durationSec?: number;
+};
+
+export type ClientDecodeOutput = {
+    durationSec: number;
+    peaks: number[];
+    extractMs: number;
+};
+
+export type DecodeToPeaksFn = (signal: AbortSignal) => Promise<ClientDecodeOutput>;
+
+export type DecodeTimings = {
+    attemptMs: number | null;
+    extractMs: number | null;
+};
+
+export type ClientDecodeResult = WaveformLoadState & {
+    isDecodeSkipped: boolean;
+    timings: DecodeTimings;
+    reason?: DecodeSkipReason;
+    error?: WaveformError;
+};
+
+export type ClientDecodeRequest = {
+    compressedBytes?: number;
+    durationSec?: number;
+    fetchArrayBuffer: (signal: AbortSignal) => Promise<ArrayBuffer>;
+    signal?: AbortSignal;
+};
+
 /**
  * Async boundary for waveform data. Implementations may fetch fixtures, decode client-side,
  * or load Conversion reps — callers only observe WaveformLoadState.
